+++ tests/h5_repack.py
@@ -1,3 +1,8 @@
+'''
+This code is not finished 
+DO NOT use it with your data!
+'''
+
 import os
 import sys
 import argparse
@@ -5,30 +10,56 @@ import shlex
 import time
 import subprocess
 import random
+import h5py
 
 
 ext = ('.h5',) # legal extensions of hdf5 file
 
+
+parser = argparse.ArgumentParser(description='Process ptrepack to shrink overwritten .h5 file size.')
+parser.add_argument('integers', metavar='path', type=str, nargs='?', help='path of a file or a folder (repack all .h5 files in the folder).')
+args = parser.parse_args()
+parser.print_help() # print the help information when run the code w/ -h or --help. Can be commonted if don't want to show
+
 # command to process the repack
 def repack_command(inpath):
+    indir = os.path.dirname(inpath)
     infilename, infileext = os.path.splitext(inpath)
     tempfilename = 'tempfile' + str(random.randrange(1e5, 9e5)) + infileext
+    temppath = os.path.join(indir, tempfilename) # use the same folder to save the tempfile
     print('repacking ' + inpath)
-    print('to ' + tempfilename + ' ...')
-    command = ["ptrepack", "-o", "--chunkshape=auto", "--propindexes", inpath, tempfilename]
-    p = subprocess.Popen(command)
+    print('to ' + temppath + ' ...')
+    print('Original file size: {}'.format(os.path.getsize(inpath)))
+
+    # ptrepack seams does not work with 
+    # command = ["ptrepack", "-o", "--chunkshape=auto", "--propindexes", inpath, temppath]
+    # p = subprocess.Popen(command)
+
+    # just use simple creating a new file and copy the data
+    with h5py.File(inpath, 'r') as fh_in:
+        with h5py.File(temppath, 'w') as fh_out:
+            fh_out.create_group('data')
+            fh_out.create_group('raw')
+            fh_out.create_group('prop')
+            
+            # copy attribute of file
+            for attr in fh_in.attrs:
+                print('attr: ', attr)
+                fh_out.attrs[attr] = fh_in.attrs[attr]
+            # copy datasets and groups
+            for key in fh_in.keys():
+                print('key: ', key)
+                fh_in.copy(key, fh_out[key], shallow=True)
+
     # delete tempfile
     print('removing original file ...')
     os.remove(inpath)
     print('saving tempfile back to ' + inpath + ' ...')
-    os.rename(tempfilename, inpath)
+    os.rename(temppath, inpath)
+    print('New file size: {}'.format(os.path.getsize(inpath)))
 
     print('done for ' + inpath)
     print('\n')
-
-parser = argparse.ArgumentParser(description='Process ptrepack to shrink overwritten .h5 file size.')
-parser.add_argument('integers', metavar='path', type=str, nargs='?', help='path of a file or a folder (repack all .h5 files in the folder).')
-parser.print_help()
 
 print('\n===============\n')
 # check the args
@@ -45,7 +76,7 @@ paths = list(filter(lambda p: (os.path.exists(p) and p.endswith(ext)) or os.path
 
 print('Input unique path(s):\n{}'.format('\n'.join(paths)))
 
-
+print('\n\n')
 if not paths: # no available path
     print('No available path. Code stoped!')
 else: